+++ engine/src/ecs/components/camera.rs
@@ -6,26 +6,26 @@ use specs::{Component, NullStorage, VecStorage};
 
 use crate::utils::*;
 
-const DEG_89: cgmath::Rad<f64> = cgmath::Rad(1.5533430342749532f64);
+const DEG_89: cgmath::Rad<f32> = cgmath::Rad(1.5533430342749532f32);
 
 #[derive(Component, Debug, Clone)]
 #[storage(VecStorage)]
 pub struct Camera {
     position: Vec3F,
-    euler_angles: cgmath::Euler<cgmath::Rad<f64>>,
-    fovy: cgmath::Rad<f64>,
-    near_distance: f64,
-    far_distance: f64,
+    euler_angles: cgmath::Euler<cgmath::Rad<f32>>,
+    fovy: cgmath::Rad<f32>,
+    near_distance: f32,
+    far_distance: f32,
 }
 
 impl Default for Camera {
     fn default() -> Self {
         Self {
-            position: Vec3F::new(0f64, 0f64, 0f64),
-            euler_angles: cgmath::Euler::new(cgmath::Rad(0f64), cgmath::Rad(0f64), cgmath::Rad(0f64)),
-            fovy: cgmath::Deg(45f64).into(),
-            near_distance: 0.1f64,
-            far_distance: 1000f64,
+            position: Vec3F::new(0f32, 0f32, 0f32),
+            euler_angles: cgmath::Euler::new(cgmath::Rad(0f32), cgmath::Rad(0f32), cgmath::Rad(0f32)),
+            fovy: cgmath::Deg(45f32).into(),
+            near_distance: 0.1f32,
+            far_distance: 1000f32,
         }
     }
 }
@@ -36,7 +36,7 @@ impl Camera {
         let euler_angles = cgmath::Euler::new(
             cgmath::Rad(facing.y.sin()),
             cgmath::Rad(facing.z.atan2(facing.x)),
-            cgmath::Rad(0f64),
+            cgmath::Rad(0f32),
         );
         let mut ret = Self::default();
         ret.position = position;
@@ -45,13 +45,13 @@ impl Camera {
     }
 
     // Camera matrices
-    pub fn projection_matrix(&self, aspect_ratio: f64) -> Mat4F {
+    pub fn projection_matrix(&self, aspect_ratio: f32) -> Mat4F {
         cgmath::perspective(self.fovy, aspect_ratio, self.near_distance, self.far_distance)
     }
 
     pub fn view_matrix(&self) -> Mat4F {
         let facing = self.front();
-        let location = cgmath::Point3::<f64>::new(self.position.x, self.position.y, self.position.z);
+        let location = cgmath::Point3::<f32>::new(self.position.x, self.position.y, self.position.z);
         Mat4F::look_at_dir(location, facing, Vec3F::unit_y())
     }
 
@@ -64,7 +64,7 @@ impl Camera {
         self.position += delta;
     }
 
-    pub fn push_rotation(&mut self, delta: cgmath::Euler<cgmath::Rad<f64>>) {
+    pub fn push_rotation(&mut self, delta: cgmath::Euler<cgmath::Rad<f32>>) {
         self.euler_angles = cgmath::Euler::new(
             self.euler_angles.x + delta.x,
             self.euler_angles.y + delta.y,
