+++ engine/src/graphics/uniform.rs
@@ -12,7 +12,7 @@ pub enum Uniform {
     Uint(u32),
     Int(i32),
     IntArray(Vec<i32>),
-    Float(f64),
+    Float(f32),
     Vec2(Vec2F),
     Vec3(Vec3F),
     Vec4(Vec4F),
